LayoutTests/platform/iphone-simulator/accessibility/tab-role-expected.txt

tab1

This test makes sure that role='tab' shows up as a distinct element from button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tab.iphoneLabel is 'AXLabel: tab1'
PASS button.iphoneTraits != tab.iphoneTraits is true
PASS tab.iphoneTraits != 0 is true
PASS tab.isSelected is true
PASS successfullyParsed is true

TEST COMPLETE
